A workflow for building releases on tag push
This should help automate building test releases for both branches
and the mainline. It takes a lot longer than building on a typical
programmer PC with fast disks and warm caches, but it will generate
reproducible builds and maybe reduce some of the by-hand work CS
contributors are doing.

The build steps are taken directly from the .bat file and
adapted to use bash.

This workflow caches the cmake `build` directory. The vcpkg
action caches the vcpkg build directory. If builds are slow,
improving caching would be the next step.